generalized confirmation dialog for reuse - implemented confirmation on deleting notifications and messages

+++ client/src/app/janusz-notifications/janusz-notifications.component.ts
@@ -7,7 +7,7 @@ import {
   trigger
 } from '@angular/animations';
 
-import { Notification, Message } from '../interfaces';
+import { Notification, Message, ConfirmationDialogData } from '../interfaces';
 import { JanuszService } from '../janusz.service';
 import { JanuszNotificationDialogComponent } from '../janusz-notification-dialog/janusz-notification-dialog.component';
 import { JanuszMessageDialogComponent } from '../janusz-message-dialog/janusz-message-dialog.component';
@@ -17,6 +17,7 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 
 import { ToastrService } from 'ngx-toastr';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-janusz-notifications',
@@ -57,6 +58,7 @@ export class JanuszNotificationsComponent implements OnInit {
     private januszService: JanuszService,
     private januszNotificationDialog: MatDialog,
     private januszMessageDialog: MatDialog,
+    private confirmDialog: MatDialog,
     private toastr: ToastrService
   ) {}
 
@@ -132,14 +134,34 @@ export class JanuszNotificationsComponent implements OnInit {
 
   onDeleteNotification(notification: Notification, event: Event) {
     event.stopPropagation();
-    this.januszService.deleteNotification(notification.id).subscribe({
-      next: () => {
-        this.toastr.success('Notification deleted');
-        this.getNotifications();
-      },
-      error: err => {
-        console.error(err);
-        this.toastr.error('Notification delete failed');
+    const data: ConfirmationDialogData = {
+      title: `Delete ${notification.label}`,
+      message: `Are you sure you want to delete ${notification.label}?`,
+      // tslint:disable-next-line: quotemark
+      confirmButton: "Yes, I'm sure",
+      cancelButton: 'Not really'
+    };
+    const dialogRef = this.confirmDialog.open(ConfirmationDialogComponent, {
+      width: '400px',
+      data,
+      autoFocus: false,
+      restoreFocus: false
+    });
+
+    dialogRef.afterClosed().subscribe({
+      next: res => {
+        if (res) {
+          this.januszService.deleteNotification(notification.id).subscribe({
+            next: () => {
+              this.toastr.success('Notification deleted');
+              this.getNotifications();
+            },
+            error: err => {
+              console.error(err);
+              this.toastr.error('Notification delete failed');
+            }
+          });
+        }
       }
     });
   }
@@ -205,14 +227,34 @@ export class JanuszNotificationsComponent implements OnInit {
   }
 
   onDeleteMessage(messageId: number) {
-    this.januszService.deleteMessage(messageId).subscribe({
-      next: () => {
-        this.toastr.success('Message deleted');
-        this.getMessages(this.expandedNotification.id);
-      },
-      error: err => {
-        console.error(err);
-        this.toastr.error('Message deletion failed');
+    const data: ConfirmationDialogData = {
+      title: `Delete message`,
+      message: `Are you sure you want to delete this message?`,
+      // tslint:disable-next-line: quotemark
+      confirmButton: "Yes, I'm sure",
+      cancelButton: 'Not really'
+    };
+    const dialogRef = this.confirmDialog.open(ConfirmationDialogComponent, {
+      width: '400px',
+      data,
+      autoFocus: false,
+      restoreFocus: false
+    });
+
+    dialogRef.afterClosed().subscribe({
+      next: res => {
+        if (res) {
+          this.januszService.deleteMessage(messageId).subscribe({
+            next: () => {
+              this.toastr.success('Message deleted');
+              this.getMessages(this.expandedNotification.id);
+            },
+            error: err => {
+              console.error(err);
+              this.toastr.error('Message deletion failed');
+            }
+          });
+        }
       }
     });
   }
